+++ app/filters/dorsale/flyboy/small_data/filter_strategy_by_task_state.rb
@@ -1,7 +1,9 @@
 class Dorsale::Flyboy::SmallData::FilterStrategyByTaskState < ::Agilibox::SmallData::FilterStrategy
+  STATES = Dorsale::Flyboy::Task::STATES + %w(on_warning_or_alert)
+
   def apply(query, value)
-    if value.in?(Dorsale::Flyboy::Task::STATES)
-      query.send(value)
+    if value.in?(STATES)
+      query.public_send(value)
     else
       query
     end